Refactor: reorganized the tree for movie-app
