+++ README.md
@@ -16,5 +16,3 @@ Made with ❤️ to programming
 - Node.js and pnpm
 
 [![Netlify Status](https://api.netlify.com/api/v1/badges/36fc89c3-8170-4f88-985b-d2b60bf46c3f/deploy-status)](https://app.netlify.com/sites/zetkolek/deploys)
-
-
