--- src/components/portfolio/movie-app/const.ts
@@ -1,4 +1,0 @@
-const getEnv = import.meta.env.PUBLIC_MOVIE_API_KEY;
-
-export const API_ENDPOINT = `https://www.omdbapi.com/?apikey=${getEnv}`;
-export const DEFAULT_ID = "tt0372784";
