+++ src/components/portfolio/movie-app/index.tsx
@@ -1,7 +1,7 @@
 import { QueryClientProvider } from "@tanstack/react-query";
-import queryClient from "@/components/portfolio/movie-app/queryClient";
+import queryClient from "@/utils/queryClient";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import AppRouting from "./components/AppRouting";
+import AppRouting from "./components/common/AppRouting";
 
 const MovieAppComponent = () => {
 	return (
