+++ src/features/movie-app/const.ts
@@ -1,0 +1,2 @@
+export const API_ENDPOINT = `https://www.omdbapi.com/`;
+export const DEFAULT_ID = "tt0372784";
