+++ src/pages/movie-app.astro
@@ -26,7 +26,7 @@ const {
 				"Playwright",
 			]}
 		/>
-		<p class="text-center pb-6">Hello there, friendly visitor from {city || "Unknown City"}!</p>
+		<p class="pb-6 text-center">Hello there, friendly visitor from {city || "Unknown City"}!</p>
 		<MoviesAppComponent client:only />
 	</ContainerSection>
 </PageLayout>
